refactor: 공통 컴포넌트 경로 별칭 반영
