+++ src/shared/components/index.ts
@@ -2,3 +2,4 @@ export { default as CommonIcon } from './Icon';
 export { default as CommonText } from './Text';
 export { default as CommonAvatar } from './Avatar';
 export { default as CommonBadge } from './Badge';
+export { default as CommonIconButton } from './IconButton';
